Adding dedicated selected property to EmbeddedShellSurfaceView to track selection status
Moved EmbeddedShellSurfaceView to dedicated .cpp/.h files, migrated from a separate deselcted signal to
a selected property that tracks selection status to be in line with how surface visibility tracking is handled.

Renamed QuickEmbeddedShellView::isCurrentView to QuickEmbeddedShellView::selected.
