+++ embeddedplatform/embeddedshellsurface.cpp
@@ -2,9 +2,9 @@
 
 #include "embeddedshellsurface.h"
 #include "embeddedshellsurface_p.h"
+#include "embeddedshellsurfaceview.h"
+#include "embeddedshellsurfaceview_p.h"
 #include <QtWaylandClient/private/qwaylandwindow_p.h>
-#include <qpa/qplatformwindow.h>
-#include <qpa/qwindowsysteminterface.h>
 
 EmbeddedShellSurface::EmbeddedShellSurface(struct ::embedded_shell_surface *shell_surface,
                                            QtWaylandClient::QWaylandWindow *window,
@@ -131,13 +131,15 @@ EmbeddedShellSurfaceView *EmbeddedShellSurface::createView(const QString &appId,
   viewPrivate->m_icon = icon;
   viewPrivate->m_sortIndex = sort_index;
 
-  connect(view, &EmbeddedShellSurfaceView::selected, d, [view, d]() {
-    if (d->m_selectedView != view) {
-      if (d->m_selectedView) {
-        d->m_selectedView->deselected();
-      }
+  connect(view, &EmbeddedShellSurfaceView::selectedChanged, d, [view, d](bool selected) {
+    if (selected) {
+      if (d->m_selectedView != view) {
+        if (d->m_selectedView) {
+          d->m_selectedView->setSelected(false);
+        }
 
-      d->m_selectedView = view;
+        d->m_selectedView = view;
+      }
     }
   });
 
@@ -190,117 +192,4 @@ void EmbeddedShellSurface::sendAppIcon(const QString &appIcon)
 {
   Q_D(EmbeddedShellSurface);
   d->set_app_icon(appIcon);
-}
-
-EmbeddedShellSurfaceViewPrivate::EmbeddedShellSurfaceViewPrivate(EmbeddedShellSurfaceView *q,
-                                                                 ::surface_view *view,
-                                                                 EmbeddedShellSurface *surf)
-    : QObject(surf)
-    , QtWayland::surface_view(view)
-    , q_ptr(q)
-    , m_sortIndex(0)
-{
-}
-
-EmbeddedShellSurfaceViewPrivate::~EmbeddedShellSurfaceViewPrivate()
-{
-  destroy();
-}
-
-EmbeddedShellSurfaceViewPrivate *EmbeddedShellSurfaceViewPrivate::get(EmbeddedShellSurfaceView *q)
-{
-  return q->d_func();
-}
-
-void EmbeddedShellSurfaceViewPrivate::surface_view_selected()
-{
-  qDebug() << __PRETTY_FUNCTION__;
-  Q_Q(EmbeddedShellSurfaceView);
-  emit q->selected();
-}
-
-EmbeddedShellSurfaceView::EmbeddedShellSurfaceView(::surface_view *view,
-                                                   EmbeddedShellSurface *surf)
-    : d_ptr(new EmbeddedShellSurfaceViewPrivate(this, view, surf))
-{
-}
-
-QString EmbeddedShellSurfaceView::appLabel() const
-{
-  Q_D(const EmbeddedShellSurfaceView);
-  return d->m_appLabel;
-}
-
-void EmbeddedShellSurfaceView::setAppLabel(const QString &appLabel)
-{
-  Q_D(EmbeddedShellSurfaceView);
-  if (d->m_appLabel == appLabel)
-    return;
-  d->m_appLabel = appLabel;
-  d->set_app_label(appLabel);
-  Q_EMIT appLabelChanged(appLabel);
-}
-
-QString EmbeddedShellSurfaceView::appIcon() const
-{
-  Q_D(const EmbeddedShellSurfaceView);
-  return d->m_appIcon;
-}
-
-void EmbeddedShellSurfaceView::setAppIcon(const QString &appIcon)
-{
-  Q_D(EmbeddedShellSurfaceView);
-  if (d->m_appIcon == appIcon)
-    return;
-  d->m_appIcon = appIcon;
-  d->set_app_icon(appIcon);
-  Q_EMIT appIconChanged(appIcon);
-}
-
-QString EmbeddedShellSurfaceView::label() const
-{
-  Q_D(const EmbeddedShellSurfaceView);
-  return d->m_label;
-}
-
-void EmbeddedShellSurfaceView::setLabel(const QString &label)
-{
-  Q_D(EmbeddedShellSurfaceView);
-  if (d->m_label == label)
-    return;
-  d->m_label = label;
-  d->set_label(label);
-  Q_EMIT labelChanged(label);
-}
-
-QString EmbeddedShellSurfaceView::icon() const
-{
-  Q_D(const EmbeddedShellSurfaceView);
-  return d->m_icon;
-}
-
-void EmbeddedShellSurfaceView::setIcon(const QString &icon)
-{
-  Q_D(EmbeddedShellSurfaceView);
-  if (d->m_icon == icon)
-    return;
-  d->m_icon = icon;
-  d->set_icon(icon);
-  Q_EMIT iconChanged(icon);
-}
-
-unsigned int EmbeddedShellSurfaceView::sortIndex() const
-{
-  Q_D(const EmbeddedShellSurfaceView);
-  return d->m_sortIndex;
-}
-
-void EmbeddedShellSurfaceView::setSortIndex(unsigned int sortIndex)
-{
-  Q_D(EmbeddedShellSurfaceView);
-  if (d->m_sortIndex == sortIndex)
-    return;
-  d->m_sortIndex = sortIndex;
-  d->set_sort_index(sortIndex);
-  Q_EMIT sortIndexChanged(sortIndex);
 }
