+++ embeddedplatform/embeddedshellsurface.h
@@ -8,16 +8,13 @@
 #include <QObject>
 
 class EmbeddedShellSurfaceView;
-class EmbeddedShellSurfaceViewPrivate;
 class EmbeddedShellSurfacePrivate;
 
 struct embedded_shell_surface;
-struct surface_view;
-class QWindow;
 
 namespace QtWaylandClient {
-class QWaylandWindow;
-class QWaylandShellSurface;
+  class QWaylandWindow;
+  class QWaylandShellSurface;
 } // namespace QtWaylandClient
 
 class Q_DECL_EXPORT EmbeddedShellSurface : public QObject
@@ -65,48 +62,6 @@ public slots:
   void sendAppId(const QString &appId);
   void sendAppLabel(const QString &appLabe);
   void sendAppIcon(const QString &appIcon);
-};
-
-class EmbeddedShellSurfaceView : public QObject
-{
-  Q_OBJECT
-  Q_DECLARE_PRIVATE(EmbeddedShellSurfaceView)
-  QScopedPointer<EmbeddedShellSurfaceViewPrivate> d_ptr;
-  Q_PROPERTY(QString appLabel READ appLabel WRITE setAppLabel NOTIFY appLabelChanged)
-  Q_PROPERTY(QString appIcon READ appIcon WRITE setAppIcon NOTIFY appIconChanged)
-  Q_PROPERTY(QString label READ label WRITE setLabel NOTIFY labelChanged)
-  Q_PROPERTY(QString icon READ icon WRITE setIcon NOTIFY iconChanged)
-  Q_PROPERTY(int sortIndex READ sortIndex WRITE setSortIndex NOTIFY sortIndexChanged)
-
-public:
-  QString appLabel() const;
-  void setAppLabel(const QString &appLabel);
-  Q_SIGNAL void appLabelChanged(const QString &appLabel);
-
-  QString appIcon() const;
-  void setAppIcon(const QString &appIcon);
-  Q_SIGNAL void appIconChanged(const QString &appIcon);
-
-  QString label() const;
-  void setLabel(const QString &label);
-  Q_SIGNAL void labelChanged(const QString &label);
-
-  QString icon() const;
-  void setIcon(const QString &icon);
-  Q_SIGNAL void iconChanged(const QString &icon);
-
-  unsigned int sortIndex() const;
-  void setSortIndex(unsigned int sortIndex);
-  Q_SIGNAL void sortIndexChanged(unsigned int sortIndex);
-
-signals:
-  void selected();
-  void deselected();
-
-private:
-  friend class EmbeddedShellSurface;
-  EmbeddedShellSurfaceView(struct ::surface_view *view,
-                           EmbeddedShellSurface *surf);
 };
 
 #endif // EMBEDDEDSHELLSURFACE_H
