+++ quickembeddedshellwindow/quickembeddedshellview.h
@@ -14,7 +14,7 @@ class EMBEDDEDSHELLWINDOW_EXPORT QuickEmbeddedShellView : public QQuickItem
   QML_NAMED_ELEMENT(View)
 
   Q_PROPERTY(QuickEmbeddedShellSurface *surface READ surface WRITE setSurface NOTIFY surfaceChanged)
-  Q_PROPERTY(bool isCurrentView READ isCurrentView NOTIFY isCurrentViewChanged)
+  Q_PROPERTY(bool selected READ selected NOTIFY selectedChanged)
 
   Q_PROPERTY(QString appId READ appId WRITE setAppId NOTIFY appIdChanged)
   Q_PROPERTY(QString appLabel READ appLabel WRITE setAppLabel NOTIFY appLabelChanged)
@@ -31,7 +31,7 @@ public:
   QuickEmbeddedShellSurface *surface() const;
   void setSurface(QuickEmbeddedShellSurface *surface);
 
-  bool isCurrentView() const;
+  bool selected() const;
 
   QString appId() const;
   void setAppId(const QString &appId);
@@ -53,7 +53,7 @@ public:
 
 signals:
   void surfaceChanged(QuickEmbeddedShellSurface *surface);
-  void isCurrentViewChanged(bool isCurrentView);
+  void selectedChanged(bool selected);
   void appIdChanged(const QString &appId);
   void appLabelChanged(const QString &appLabel);
   void appIconChanged(const QString &appIcon);
@@ -62,11 +62,11 @@ signals:
   void sortIndexChanged(quint32 sortIndex);
 
 private:
-  void setIsCurrentView(bool isCurrentView);
+  void setSelected(bool selected);
   void createView();
 
   QuickEmbeddedShellSurface *m_surface;
-  bool m_isCurrentView;
+  bool m_selected;
 
   QString m_appId;
   QString m_appLabel;
